+++ lib/goodmap.dart
@@ -22,7 +22,12 @@ export 'src/globe/globe_overlays.dart' show GlobeArc;
 export 'src/globe/good_globe.dart' show GoodGlobe;
 export 'src/globe/good_map_globe.dart' show GoodMapGlobe;
 export 'src/good_map.dart' show GoodMap;
-export 'src/lines/region.dart' show GoodMapRegionOptions, GoodMapRegionPolygon;
+export 'src/lines/region.dart'
+    show
+        GoodMapRegionOptions,
+        GoodMapRegionPolygon,
+        GoodMapRegionState,
+        GoodMapRegionStateChanged;
 export 'src/theme/basemaps.dart' show Basemaps;
 export 'src/theme/carto_basemap_config.dart'
     show GoodBasemapConfig, GoodBasemapConfigurationException;